pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "todoist-organizer"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "todoist-api-python>=3.1.0,<4",
    "anthropic>=0.40.0,<1",
    "python-dotenv>=1.0.0,<2",
]

[project.scripts]
todoist-organizer = "todoist_organizer.cli:main"
